+++ projects/_01_ingest/cqc_api/jobs/delta_clean_cqc_location_data.py
@@ -85,6 +85,7 @@ ons_cols_to_import = [
 def main(
     cqc_location_source: str,
     cleaned_ons_postcode_directory_source: str,
+    manual_postcode_corrections_source: str,
     cleaned_cqc_location_destination: str,
     gac_service_destination: str,
     regulated_activities_destination: str,
@@ -230,6 +231,7 @@ def main(
         ons_postcode_directory_df,
         postcode_matching_destination,
         dimension_update_date,
+        manual_postcode_corrections_source,
     )
 
     cqc_location_df = cqc_location_df.drop(CQCL.postal_code, CQCL.postal_address_line1)
@@ -254,6 +256,7 @@ def create_postcode_matching_dimension(
     postcode_df: DataFrame,
     dimension_location: str,
     dimension_update_date: str,
+    manual_postcode_corrections_source: str,
 ) -> DataFrame:
     """
     Creates (or updates) the postcode matching dimension by comparing current CQC data
@@ -270,6 +273,7 @@ def create_postcode_matching_dimension(
         dimension_location (str): S3 path where the dimension data should be stored
             dimension parquet file is stored.
         dimension_update_date (str): Date string in 'YYYYMMDD' format for creating partition keys.
+        manual_postcode_corrections_source (str): The s3 URI for the incorrect postcode csv.
 
     Returns:
         DataFrame: Dataframe of delta dimension table, with rows of the changes since the last update.
@@ -292,6 +296,7 @@ def create_postcode_matching_dimension(
             Keys.import_date,
         ),
         postcode_df,
+        manual_postcode_corrections_source,
     )
 
     if previous_dimension:
@@ -845,6 +850,7 @@ if __name__ == "__main__":
     (
         cqc_location_source,
         cleaned_ons_postcode_directory_source,
+        manual_postcode_corrections_source,
         cleaned_cqc_location_destination,
         gac_service_dest,
         regulated_activities_dest,
@@ -858,6 +864,10 @@ if __name__ == "__main__":
         (
             "--cleaned_ons_postcode_directory_source",
             "Source s3 directory for parquet ONS postcode directory dataset",
+        ),
+        (
+            "--manual_postcode_corrections_source",
+            "Source s3 location for incorrect postcode csv dataset",
         ),
         (
             "--cleaned_cqc_location_destination",
@@ -883,6 +893,7 @@ if __name__ == "__main__":
     main(
         cqc_location_source,
         cleaned_ons_postcode_directory_source,
+        manual_postcode_corrections_source,
         cleaned_cqc_location_destination,
         gac_service_dest,
         regulated_activities_dest,